requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
rembg[cpu]>=2.0.50
pyinstaller>=6.0.0
